QUAD-143: Splitting ConceptCollectionProjectController and DictionaryProjectController into multiple controllers, one each to list, delete and add.
