+++ Quadriga/src/main/java/edu/asu/spring/quadriga/web/workbench/AddProjectDictionaryController.java
@@ -18,7 +18,6 @@ import edu.asu.spring.quadriga.aspects.annotations.CheckedElementType;
 import edu.asu.spring.quadriga.aspects.annotations.ElementAccessPolicy;
 import edu.asu.spring.quadriga.domain.dictionary.IDictionary;
 import edu.asu.spring.quadriga.domain.workbench.IProject;
-import edu.asu.spring.quadriga.domain.workbench.IProjectDictionary;
 import edu.asu.spring.quadriga.exceptions.QuadrigaAccessException;
 import edu.asu.spring.quadriga.exceptions.QuadrigaStorageException;
 import edu.asu.spring.quadriga.service.dictionary.IDictionaryManager;
@@ -27,7 +26,7 @@ import edu.asu.spring.quadriga.service.workbench.IRetrieveProjectManager;
 import edu.asu.spring.quadriga.web.login.RoleNames;
 
 @Controller
-public class DictionaryProjectController {
+public class AddProjectDictionaryController {
 
     @Autowired
     IRetrieveProjectManager projectManager;
@@ -67,54 +66,8 @@ public class DictionaryProjectController {
             projectDictionaryManager.addDictionaryToProject(projectid, values[i], userId);
         }
         attr.addFlashAttribute("show_success_alert", true);
-        attr.addFlashAttribute("success_alert_msg", "Dictionaries added to workspace successfully.");
+        attr.addFlashAttribute("success_alert_msg", "Dictionaries added to project successfully.");
         return "redirect:/auth/workbench/projects/" + projectid;
     }
 
-    @RequestMapping(value = "auth/workbench/{projectid}/dictionaries", method = RequestMethod.GET)
-    public String listProjectDictionary(HttpServletRequest req, @PathVariable("projectid") String projectid,
-            Model model, Principal principal) throws QuadrigaStorageException {
-        String userId = principal.getName();
-        List<IProjectDictionary> dicitonaryList = projectDictionaryManager.listProjectDictionary(projectid, userId);
-        model.addAttribute("dicitonaryList", dicitonaryList);
-        IProject project = projectManager.getProjectDetails(projectid);
-        model.addAttribute("project", project);
-        return "auth/workbench/project/dictionaries";
-    }
-
-    @AccessPolicies({ @ElementAccessPolicy(type = CheckedElementType.PROJECT, paramIndex = 1, userRole = {
-            RoleNames.ROLE_COLLABORATOR_ADMIN, RoleNames.ROLE_PROJ_COLLABORATOR_ADMIN }) })
-    @RequestMapping(value = "auth/workbench/{projectid}/deletedictionary", method = RequestMethod.GET)
-    public String deleteProjectDictionary(@PathVariable("projectid") String projectid, Model model, Principal principal)
-            throws QuadrigaStorageException, QuadrigaAccessException {
-        String userId = principal.getName();
-
-        List<IProjectDictionary> dicitonaryList = projectDictionaryManager.listProjectDictionary(projectid, userId);
-        model.addAttribute("dicitonaryList", dicitonaryList);
-        IProject project = projectManager.getProjectDetails(projectid);
-        model.addAttribute("project", project);
-        return "auth/workbench/project/deletedictionaries";
-    }
-
-    @AccessPolicies({ @ElementAccessPolicy(type = CheckedElementType.PROJECT, paramIndex = 2, userRole = {
-            RoleNames.ROLE_COLLABORATOR_ADMIN, RoleNames.ROLE_PROJ_COLLABORATOR_ADMIN }) })
-    @RequestMapping(value = "auth/workbench/{projectid}/deletedictionaries", method = RequestMethod.POST)
-    public String deleteProjectDictionary(HttpServletRequest req, @PathVariable("projectid") String projectid,
-            Model model, Principal principal, RedirectAttributes attr)
-                    throws QuadrigaStorageException, QuadrigaAccessException {
-        String userId = principal.getName();
-
-        String[] values = req.getParameterValues("selected");
-        if (values == null) {
-            attr.addFlashAttribute("show_error_alert", true);
-            attr.addFlashAttribute("error_alert_msg", "Please select a Dictionary.");
-            return "redirect:/auth/workbench/" + projectid + "/deletedictionary";
-        }
-        for (int i = 0; i < values.length; i++) {
-            projectDictionaryManager.deleteProjectDictionary(projectid, userId, values[i]);
-        }
-        attr.addFlashAttribute("show_success_alert", true);
-        attr.addFlashAttribute("success_alert_msg", "Dictionaries deleted from workspace successfully.");
-        return "redirect:/auth/workbench/projects/" + projectid;
-    }
 }
